app/requirements.txt
fastapi
asyncpg
sqlalchemy[asyncio]
alembic
psycopg2-binary
uvicorn[standard]
python-dotenv
